requirements.txt
Flask==2.0.1
PyJWT==2.4.0
redis==4.0.2
Werkzeug==2.0.3  # Pin Werkzeug to a version compatible with Flask 2.0.1